Start IOT_Home_Service Project
Because I was always interested in changing everyday life, IOT was
the best fit for such a concern.
For this reason, I needed a program to help me, and I decided to created
a IOT project right now.

This project aims to provide information about the environment through
sensors. It is also intended to provide information that is not through
the sensor.

Please see the Readme for more information on the project.