Make mqtt.py for sensor
Model of Django is represented by python class. Then, Django provides
a Database Access API through these model classes. To summarize, the
sensor values are stored in the database, and model must represent the
table about that sensor as a class so that Django can access the sensors
stored in that database.

mqtt.py takes the message (sensor value) of the topic published to the
mqtt broker and performs the operation described above for that value