Implement WebSocket, Social Authentication
Implement WebSocket with 'JavaScript' and 'Node.js'. Because you need to
update sensor value from mqtt to web client realtime. And later, you
need WebSocket to implement the chat system. Of course you can use the
library for django without using Node.js, but recommend Node.js because
Node.js is easier to use with WebSocket.

Social Authentication is for convenience, control and OpenAPI. One of
goals of this project is the Smart Mirror Mode, which controled by user.
So user account should always remember whether the user is in Smart-
Mirror mode or not. In addition, you should use OpenApi to provide many
useful features that will benefit you. Each user needs different
function. So you have to use social authentication to provide this
separately.