Implement Worker&Master, and bulid a base of API.
Implement Worker & Master System. The Worker mainly process data for a
specific user such as a sensor, an emial, a calendar, etc., and receives
and processes requests from a user. Master is you! The master, client,
and user are all the same. If users who use this service connect at the
same time, they should be separated. Therefore, the Worker & Master
System is a system that divide various users and provides each user
with necessary data.

Build an API foundation. In some api cases, data can be accessed
through authorization(like email), and in some api cases, data can only
be accessed using an api key(like Weather). Therefore, you need to
build an API base for these cases.

This commit was to build Worker & Master system and api foundation.
In addition, the direction of the desing of UI was decided by
'Material Design'. Based on this direction, I implemented some UI.
I improved the intuitiveness of the code by modularizing js and css,and
used library of 'https://getmdl.io/' for css.