Implement Email System
The Email System is one of the core in API services. Because email is
still a very important communication tool, users should be able to
easily receive email information. Because the core of this email system
is 'receiving', the implementation of 'sending' is not yet in order.
By default, the emails of the user's email account are fetched and
stored in db. At this time, the composition of the email is only
'Title', 'From' and 'Date', but 'Body' is excluded. This is to reduce
the capacity and the time it takes to fetch it in at least a few secon-
ds. One way to reduce the amount of time currently in use is to use
'thread' to fetch the emails. The excluded 'Body' fetched when the user
requests it.

This commit implements the Email System. Soon I will add real-time
notification via idle mode.