Implement a weather api and schedule system
Implement a weather api. Use api-key of 'https://www.wunderground.com'.

Implement schedule system. The schedule system is used by the 'worker'
to perform worker's tasks on a regular basis and transmit the data to
master. When a new master is connected, the schedule for the master is
initially set. Typical tasks in the schedule are organizing information
that should ne periodically to users, such as mail polling or weather
information.

This commit was to built schedule system. In addition, made ui about
the weather, and supplemented other parts.